Created project functions and tests

+++ app/Http/Controllers/ProjectController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Project;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
 
 class ProjectController extends Controller
 {
 
-    public function create(User $user, $name, $description){
+    public function create($user, $name, $description){
 
         $project = Project::create([
             'name' => $name,
@@ -22,10 +24,12 @@ class ProjectController extends Controller
 
     }
 
-    public function update(User $user, Project $project, $name = null, $description = null){
+    public function update($user, $project_id, $name = null, $description = null){
+
+        $project = Project::find($project_id);
 
         if ($user->cannot('update', $project)) {
-            abort(403);
+            abort(403, 'You are not authorized to perform this action.');
         }
 
         $project->update([
@@ -38,10 +42,18 @@ class ProjectController extends Controller
 
     }
 
-    public function delete(User $user, Project $project){
+    public function delete($user, $project_id){
+
+        $project = Project::find($project_id);
 
         if ($user->cannot('delete', $project)) {
-            abort(403);
+            abort(403, 'You are not authorized to perform this action.');
+        }
+
+        //if project has contributions, abort
+
+        if ($project->contributions()->count() > 0){
+            abort(403, 'You cannot delete a project that has contributions.');
         }
 
         $project->delete();
